Unnecessary Elements Deleted, Extension Methods Utilized
I have deleted the "create" link that threw an exception and that I have left intact during the past stage of this test.

I have also corrected the generation of the ProfileLink property for DetailsViewModel as it was being set as the same link for the team logo.

I have also moved some the converting functions from the team class into the TestExtensions class which I have just created.