All Requirements Met
Now the user is being redirected to the listing page after having editted a club. An alert (just a div with text) is displayed whenever a user does this, indicating to which club they have made the change.

With this I believe I have met all the requirements for this test.